Title: 자주 등장한 top k 숫자, Time: 66ms, Memory: 4MB, Status: Passed - CodeTree

+++ 240405/자주 등장한 top k 숫자/top-k-frequent-elements.cpp
@@ -22,7 +22,7 @@ int main() {
 
 	sort(temp.begin(), temp.end());
 
-	for (int i = temp.size() - 1; i >= temp.size() - k; i--) {
+	for (int i = (int)temp.size() - 1; i >= (int)temp.size() - k; i--) {
 		cout << temp[i].second << " ";
 	}
 }
